SPOI-7044: Create Expression Evaluator support Java Expression Language
The expresion are based on java expressions except the operands needs to be registered before and put inside ${}.

The expression evaluator gives options to evaluate a complete function as well as an expression.

The evaluator provides with some basic utilities to be used directly into expression and also give freedom to add a custom method to be registered with ExpressionEvaluator.
This custom method once registered becomes callable into expression directly.